extracted worker-queue into pckg
